# mint-ensemble-manager/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Mint Ensemble Manager API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/v1",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "oauth2": {
        "type": "oauth2",
        "flows": {
          "implicit": {
            "authorizationUrl": "https://portals.tapis.io/v3/oauth2/authorize",
            "scopes": {}
          }
        }
      }
    },
    "schemas": {
...
      "ModelThread": {
        "description": "",
        "required": [
          "thread_id"
        ],
        "type": "object",
        "properties": {
          "thread_id": {
            "description": "The Modeling thread id",
            "type": "string"
          },
          "model_id": {
            "description": "The Model id (optional. Set to null to run all)",
            "type": "string"
          }
        }
      },
...
    }
  },
  "paths": {
    "/executionQueue": {
      "get": {
        "summary": "Get the current execution queue",
        "description": "Retrieves the current state of the execution queue",
        "tags": [
          "Execution Queue"
        ],
        "responses": {
          "202": {
            "description": "Successfully retrieved execution queue",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "success"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Default error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Empty the execution queue",
        "description": "Removes all items from the execution queue",
        "tags": [
          "Execution Queue"
        ],
        "responses": {
          "202": {
            "description": "Successfully emptied execution queue",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "success"
                    }
                  }
                }
              }
            }
          },
          "406": {
            "description": "Error emptying execution queue"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/executions": {
      "post": {
        "summary": "Submit modeling thread for execution.",
        "operationId": "submitExecution",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "oauth2": []
          }
        ],
        "requestBody": {
          "description": "Modeling thread scenario/subgoal/id",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelThread"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful response"
          },
          "default": {
            "description": "An error occurred"
          }
        }
      }
    },
    "/executions/{executionId}/logs": {
      "get": {
        "summary": "Get Execution Logs",
        "description": "Get the logs of an execution.",
        "tags": [
          "Executions"
        ],
        "security": [
          {
            "BearerAuth": []
          },
          {
            "oauth2": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "executionId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "9bc5bbfb-d76c-4d0b-87cc-f89e945a062e-007"
          }
        ],
        "responses": {
          "200": {
            "description": "Execution Logs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "success"
                    },
                    "message": {
                      "type": "string",
                      "description": "The log content or success message",
                      "example": "Execution logs fetched successfully"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Default error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "error"
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/logs": {
      "get": {
        "summary": "Fetch logs for an execution.",
        "operationId": "fetchLog",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "oauth2": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "ensemble_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Log Details",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "Log fetched successfully"
                }
              }
            }
          },
          "default": {
            "description": "Default error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "error"
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/executionEngines/tapis": {
      "post": {
        "summary": "Submit modeling thread for execution using Tapis.",
        "operationId": "submitTapisExecution",
        "tags": [
          "Execution Engine"
        ],
        "requestBody": {
          "description": "Modeling thread",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelThread"
              },
              "example": {
                "thread_id": "108bguHTBBNLlZaPMLlM",
                "model_id": "https://w3id.org/okn/i/mint/modflow_2005_BartonSprings_avg"
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": [],
            "oauth2": []
          }
        ],
        "responses": {
          "202": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Execution submitted successfully"
                    },
                    "job_ids": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "example": "1234567890"
                      }
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Default error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "error"
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}